+++ src/lightly_train/_commands/train.py
@@ -103,7 +103,7 @@ def train(
 ) -> None:
     """Train a self-supervised model.
 
-    See the documentation for more information: https://docs.lightly.ai/train/stable/train.html
+    See the documentation for more information: https://docs.lightly.ai/train/stable/pretrain_distill.html
 
     The training process can be monitored with TensorBoard:
 
@@ -157,7 +157,7 @@ def train(
 
             If you want to resume training from an interrupted or crashed run, use the
             ``resume_interrupted`` parameter instead.
-            See https://docs.lightly.ai/train/stable/train/index.html#resume-training
+            See https://docs.lightly.ai/train/stable/pretrain_distill/index.html#resume-training
             for more information.
         resume_interrupted:
             Set this to True if you want to resume training from an **interrupted or
@@ -170,7 +170,7 @@ def train(
 
             If you want to further pretrain a model or change the training parameters,
             use the ``checkpoint`` parameter instead.
-            See https://docs.lightly.ai/train/stable/train/index.html#resume-training
+            See https://docs.lightly.ai/train/stable/pretrain_distill/index.html#resume-training
             for more information.
         overwrite:
             Overwrite the output directory if it already exists. Warning, this might