+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Fix rank-dependent issues with `num_denoising` in `train_object_detection` on multi-GPU setups.
+- Add support for specifying data configs in YAML format.
 
 ### Changed