+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Improve the layout of logging of object detection.
+- Use `lightly-train.pretrain()` as the interface of pretraining and distilation instead of `lightly-train.train()`. The `lightly-train.train()` interface will be
+  deprecated in the future.
 
 ### Deprecated