+++ CHANGELOG.md
@@ -17,6 +17,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Deprecated
 
+- Deprecate `reuse_class_head` argument in `train` command. The model will now
+  automatically reuse the classification head only when the number of classes in the
+  data config matches that in the checkpoint. Otherwise, the classification head will
+  be re-initialized.
+
 ### Removed
 
 ### Fixed